+++ src/Ydb.Sdk/src/GrpcRequestSettings.cs
@@ -9,7 +9,7 @@ public class GrpcRequestSettings
     internal static readonly GrpcRequestSettings DefaultInstance = new();
 
     public string TraceId { get; set; } = string.Empty;
-    public TimeSpan? TransportTimeout { get; set; }
+    public TimeSpan TransportTimeout { get; set; } = TimeSpan.Zero;
     public ImmutableArray<string> CustomClientHeaders { get; } = new();
 
     internal long NodeId { get; set; }